+++ noodles-bcf/src/writer/vcf_record/site/info.rs
@@ -22,8 +22,8 @@ pub fn write_info<W>(
 where
     W: Write,
 {
-    for field in info.as_ref().values() {
-        write_info_field(writer, string_string_map, field.key(), field.value())?;
+    for (key, value) in info.as_ref() {
+        write_info_field(writer, string_string_map, key, value.as_ref())?;
     }
 
     Ok(())