+++ noodles-vcf/CHANGELOG.md
@@ -23,6 +23,9 @@
   * vcf/header/record/value/map/info: Change conversion from key value to a
     reference (`&info::Key`).
 
+  * vcf/record/info: Change underlying map to `Key`-`Value` rather
+    than `Key`-`Field`.
+
 ## 0.23.0 - 2022-11-29
 
 ### Added